Implement Agentic Unified Omni-Channel Inbox UI
