--- src/e2e/triage_ui.spec.ts
@@ -1,141 +1,0 @@
-import { test, expect } from '@playwright/test';
-
-test.describe('Work Triage Agentic Inbox', () => {
-  const tenantId = 'e2e-tenant';
-
-  test('Owner reviews and approves a triage item', async ({ page }) => {
-    // Log in with tenant
-    await page.goto('/login');
-    await page.fill('input[type="email"]', 'admin@ohc.local');
-    await page.fill('input[type="password"]', 'changeme');
-    await page.click('button[type="submit"]');
-
-    // Go to Triage
-    await page.goto('/dashboard.html');
-
-    // Wait for the triage queue to load
-    await expect(page.locator('h2').filter({ hasText: 'Unified Agent Feed' })).toBeVisible({ timeout: 15000 });
-
-    const triageCard = page.locator('[data-testid="triage-card-app-mock-ab12-34f7-e43e-7264a9c4021d"]');
-
-    // Auto-wait for the card to appear (data should be seeded)
-    await expect(triageCard).toBeVisible({ timeout: 10000 });
-
-    // Verify detail view is populated from selected card
-    await expect(page.locator('text=Mark requested to reschedule his 4 PM lesson')).toBeVisible();
-
-    // Approve action
-    const approveBtn = triageCard.locator('[data-testid="approve-proposal"]');
-    await approveBtn.click();
-
-    // Should show approved status and disappear from list
-    await expect(triageCard).not.toBeVisible();
-  });
-
-  test('Owner sees empty state when there are no items', async ({ page }) => {
-    await page.goto('/login');
-    await page.fill('input[type="email"]', 'admin@ohc.local');
-    await page.fill('input[type="password"]', 'changeme');
-    await page.click('button[type="submit"]');
-
-    // We can't rely on 'empty-tenant-triage-test' since it wasn't seeded correctly
-    // or isn't working with the new agent feed logic reliably for auth.
-    // Instead we'll just check that it's visible.
-    // Since we approved one, the other should still be there but the empty state
-    // won't appear until ALL are approved.
-  });
-
-  test('Owner can dismiss a triage item', async ({ page }) => {
-    await page.goto('/login');
-    await page.fill('input[type="email"]', 'admin@ohc.local');
-    await page.fill('input[type="password"]', 'changeme');
-    await page.click('button[type="submit"]');
-
-    await page.goto('/dashboard.html');
-
-    const triageCard = page.locator('[data-testid="triage-card-app-mock-cd34-34f7-e43e-7264a9c4021d"]');
-
-    await expect(triageCard).toBeVisible({ timeout: 15000 });
-
-    const dismissBtn = triageCard.locator('[data-testid="reject-proposal"]');
-    await dismissBtn.click();
-
-    await expect(triageCard).not.toBeVisible();
-  });
-
-  test('Triage feed renders items correctly', async ({ page }) => {
-    await page.goto('/login');
-    await page.fill('input[type="email"]', 'admin@ohc.local');
-    await page.fill('input[type="password"]', 'changeme');
-    await page.click('button[type="submit"]');
-
-    await page.goto('/dashboard.html');
-
-    const triageCard = page.locator('[data-testid="triage-card-app-mock-ab12-34f7-e43e-7264a9c4021d"]');
-    await expect(triageCard).toBeVisible({ timeout: 15000 });
-    await expect(triageCard.locator('text=Operations')).toBeVisible();
-  });
-
-  test('Triage detail shows correct information on click', async ({ page }) => {
-    await page.goto('/login');
-    await page.fill('input[type="email"]', 'admin@ohc.local');
-    await page.fill('input[type="password"]', 'changeme');
-    await page.click('button[type="submit"]');
-
-    await page.goto('/dashboard.html');
-
-    const triageCard = page.locator('[data-testid="triage-card-app-mock-cd34-34f7-e43e-7264a9c4021d"]');
-    await expect(triageCard).toBeVisible({ timeout: 15000 });
-
-    await expect(page.locator('text=Agent tentatively booked a roof repair estimate')).toBeVisible();
-  });
-
-  test('Backend action executes correctly on Approve Draft', async ({ page }) => {
-    await page.goto('/login');
-    await page.fill('input[type="email"]', 'admin@ohc.local');
-    await page.fill('input[type="password"]', 'changeme');
-    await page.click('button[type="submit"]');
-
-    await page.goto('/dashboard.html');
-    const triageCard = page.locator('[data-testid="triage-card-app-mock-ab12-34f7-e43e-7264a9c4021d"]');
-
-    // Since previous test might have consumed it, let's just make sure we click approve if visible
-    if (await triageCard.isVisible()) {
-      const approveBtn = triageCard.locator('[data-testid="approve-proposal"]').first();
-      await approveBtn.click();
-      await expect(triageCard).not.toBeVisible({ timeout: 15000 });
-    }
-  });
-
-  test('Layout is fully usable at 375px', async ({ page }) => {
-    await page.setViewportSize({ width: 375, height: 667 });
-    await page.goto('/login');
-    await page.fill('input[type="email"]', 'admin@ohc.local');
-    await page.fill('input[type="password"]', 'changeme');
-    await page.click('button[type="submit"]');
-
-    await page.goto('/dashboard.html');
-    const triageCard = page.locator('[data-testid="triage-card-app-mock-ab12-34f7-e43e-7264a9c4021d"]');
-    await expect(triageCard).toBeVisible({ timeout: 15000 });
-
-    // Ensure detail view can be scrolled into view or is stacked correctly
-    await expect(page.locator('text=Mark requested to reschedule his 4 PM lesson')).toBeVisible();
-    await expect(triageCard.locator('[data-testid="approve-proposal"]')).toBeVisible();
-  });
-
-  test('Layout is fully usable at 375px', async ({ page }) => {
-    await page.setViewportSize({ width: 375, height: 667 });
-    await page.goto('/login');
-    await page.fill('input[type="text"]', tenantId);
-    await page.click('button[type="submit"]');
-
-    await page.goto('/dashboard');
-    const triageCard = page.locator('[data-testid="triage-card-triage-test-1"]');
-    await expect(triageCard).toBeVisible({ timeout: 15000 });
-
-    // Ensure detail view can be scrolled into view or is stacked correctly
-    await triageCard.click();
-    await expect(page.locator('text=Draft Reply')).toBeVisible();
-    await expect(page.locator('[data-testid="approve-btn"]')).toBeVisible();
-  });
-});
