+++ src/ui/tauri/src/ui/dashboard.html
@@ -740,6 +740,7 @@
       <div class="ohc-growth-card" id="triage-section" style="display: none; padding: 20px; border: 1px solid rgba(255, 255, 255, 0.4); border-radius: 16px;">
         <h2 style="font-size: 24px; font-weight: 700; color: #1D1D1F;">Unified Agent Feed</h2>
         <p style="color: #6b7280; margin-bottom: 16px;">Review AI-prepared actions and reply drafts across all channels.</p>
+        <div style="display: flex; align-items: center; justify-content: space-between; margin-bottom: 16px;"><h2 style="margin: 0; font-size: 20px; font-weight: 700;">Needs Attention Today</h2></div>
         <div class="triage-tab-container">
           <div class="triage-tab active" id="tab-proposals" onclick="switchTriageTab('proposals')">Proposals</div>
           <div class="triage-tab" id="tab-activity" onclick="switchTriageTab('activity')">Activity Feed</div>
@@ -1389,13 +1390,13 @@
         triageSection.style.display = 'block';
         try {
           const tenantId = localStorage.getItem('tenant_id') || localStorage.getItem('tenant') || 'default';
-          const res = await fetch(`/api/agent-feed?tenant_id=${encodeURIComponent(tenantId)}`, {
+          const res = await fetch(`/api/ui/triage?tenant_id=${encodeURIComponent(tenantId)}`, {
             headers: { 'X-Tenant-ID': tenantId, 'X-User-ID': 'default' }
           });
           if (!res.ok) throw new Error("Failed to load");
 
           const data = await res.json();
-          agentFeedItems = data.items || [];
+          agentFeedItems = Array.isArray(data) ? data : data.items || [];
           renderTriageItems(agentFeedItems);
         } catch (err) {
           console.error("Agent Feed load error:", err);
@@ -1406,7 +1407,7 @@
 
       function renderTriageItems(items) {
         if (currentTriageTab === 'proposals') {
-          const proposals = items.filter(i => i.lifecycle_state !== 'APPROVED' && i.lifecycle_state !== 'DISMISSED');
+          const proposals = items;
           if (proposals.length === 0) {
             triageQueue.innerHTML = '<div class="triage-card empty">No items need your attention right now. Great job!</div>';
             return;
@@ -1479,6 +1480,31 @@
                 <div class="triage-controls" style="display: flex; gap: 8px;">
                   <button class="triage-btn-approve" data-testid="approve-instagram-dm" onclick="handleTriageAction('${item.id}', 'APPROVED')" style="flex: 1; background: #0066FF; color: white; padding: 12px; border-radius: 12px; font-weight: 600; border: none; cursor: pointer;">Approve Reply</button>
                   <button class="triage-btn-dismiss" onclick="handleTriageAction('${item.id}', 'DISMISSED')" style="flex: 1; background: rgba(0,0,0,0.05); color: #1d1d1f; padding: 12px; border-radius: 12px; font-weight: 600; border: none; cursor: pointer;">Edit Draft</button>
+                </div>
+              </div>
+            `;
+          }
+
+          if (item.source === 'Instagram DM' || item.source === 'WhatsApp' || item.source === 'Email') {
+            let payloadHtml = "";
+            if (item.action_type) {
+              payloadHtml = `<div style="background: rgba(0,0,0,0.03); padding: 12px; border-radius: 12px; margin-bottom: 16px; font-size: 14px; color: #333; font-style: italic;">${item.action_payload || ""}</div>`;
+            }
+            return `
+              <div class="triage-item" data-testid="triage-card-${item.id}" id="triage-${item.id}" style="background: rgba(255, 255, 255, 0.4); backdrop-filter: blur(20px); border-radius: 16px; padding: 16px; border: 1px solid rgba(255,255,255,0.6);">
+                <div class="triage-header" style="margin-bottom: 12px;">
+                  <div class="triage-source" style="display: flex; align-items: center; gap: 8px;">
+                    <span style="font-size: 18px;">💬</span> <strong>${item.source}</strong>
+                  </div>
+                  <div class="triage-priority">${item.priority || "Normal"}</div>
+                </div>
+                <div class="triage-context" style="font-weight: 600; margin-bottom: 12px; font-size: 15px;">
+                  ${item.context || "New message"}
+                </div>
+                ${payloadHtml}
+                <div class="triage-controls" style="display: flex; gap: 8px;">
+                  <button class="triage-btn-approve" data-testid="approve-btn" onclick="handleTriageAction('${item.id}', true)" style="flex: 1; background: #0066FF; color: white; padding: 12px; border-radius: 12px; font-weight: 600; border: none; cursor: pointer;">Approve Reply</button>
+                  <button class="triage-btn-dismiss" data-testid="reject-proposal" onclick="handleTriageAction('${item.id}', false)" style="flex: 1; background: rgba(0,0,0,0.05); color: #1d1d1f; padding: 12px; border-radius: 12px; font-weight: 600; border: none; cursor: pointer;">Edit Draft</button>
                 </div>
               </div>
             `;
@@ -1570,12 +1596,13 @@
         if (itemEl) itemEl.style.opacity = '0.5';
 
         try {
-          let url = `/api/agent-feed/${id}/state`;
+          let url = `/api/ui/triage/action`;
+          let payload = { triage_item_id: id, approved: state === true || state === "APPROVED" };
           const tenantId = localStorage.getItem('tenant_id') || localStorage.getItem('tenant') || 'default';
           const res = await fetch(`${url}?tenant_id=${encodeURIComponent(tenantId)}`, {
-            method: 'PUT',
+            method: 'POST',
             headers: { 'Content-Type': 'application/json', 'X-Tenant-ID': tenantId, 'X-User-ID': 'default' },
-            body: JSON.stringify({ state })
+            body: JSON.stringify(payload)
           });
 
           if (!res.ok) throw new Error("Update failed");
